Improvement: add multi-arch support for lib-xclient-area
git-svn-id: file:///home/code/svn/ovd/trunk@5091 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ client/java/libXClientArea/src/org_ulteo_utils_jni_WorkArea.c
@@ -67,7 +67,7 @@ Rectangle* getWorkAreaRect() {
 
 	rect = malloc(sizeof(Rectangle));
 	if (rect == NULL) {
-		fprintf(stderr, "Not enough memory: need %lu bytes", sizeof(Rectangle));
+		fprintf(stderr, "Not enough memory: need %u bytes", (unsigned int)(sizeof(Rectangle)));
 		return NULL;
 	}
